+++ src/main/java/com/daniel/core/domain/repository/IInvestmentTypeRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface IInvestmentTypeRepository {
     List<InvestmentType> listAll();
     void save(String name);
-    void rename(int id, String newName);
+    void rename(long id, String newName);
     void delete(long id);
 }
